move database back into iroh_bytes
