+++ iroh-bytes/Cargo.toml
@@ -26,6 +26,7 @@ iroh-net = { path = "../iroh-net" }
 multibase = "0.9.1"
 num_cpus = "1.15.0"
 once_cell = "1.17.0"
+portable-atomic = "1.4.0"
 postcard = { version = "1", default-features = false, features = ["alloc", "use-std", "experimental-derive"] }
 quinn = "0.10"
 rand = "0.8"
@@ -41,11 +42,14 @@ tokio-util = { version = "0.7", features = ["io-util", "io", "rt"] }
 tracing = "0.1"
 tracing-futures = "0.2.5"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+walkdir = "2.3.3"
 
 [dev-dependencies]
 http-body = "0.4.5"
 proptest = "1.0.0"
 rand = "0.8"
+tempfile = "3.6.0"
+testdir = "0.8.0"
 tokio = { version = "1", features = ["macros", "test-util"] }
 
 [features]
