+++ iroh-bytes/src/database.rs
@@ -1,0 +1,7 @@
+//! The concrete database used by the iroh binary.
+pub mod in_mem;
+pub use in_mem::*;
+mod blobs;
+pub use blobs::*;
+pub mod flat;
+pub use flat::*;
