+++ iroh-bytes/src/lib.rs
@@ -4,6 +4,7 @@
 #![recursion_limit = "256"]
 
 pub mod collection;
+pub mod database;
 pub mod get;
 pub mod protocol;
 pub mod provider;
