+++ iroh-bytes/src/util.rs
@@ -5,6 +5,7 @@ use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 use std::{fmt, result, str::FromStr};
 use thiserror::Error;
 pub mod io;
+pub mod progress;
 pub mod runtime;
 
 /// Hash type used throught.
