+++ iroh/src/commands.rs
@@ -6,8 +6,8 @@ use std::{net::SocketAddr, path::PathBuf};
 
 use anyhow::{Context, Result};
 use clap::{Parser, Subcommand};
+use iroh::bytes::{protocol::RequestToken, provider::Ticket, util::runtime, Hash};
 use iroh::rpc_protocol::*;
-use iroh_bytes::{protocol::RequestToken, provider::Ticket, util::runtime, Hash};
 use iroh_net::tls::{Keypair, PeerId};
 use quic_rpc::transport::quinn::QuinnConnection;
 use quic_rpc::RpcClient;
@@ -78,7 +78,7 @@ impl Cli {
                 } else if let (Some(peer), Some(hash)) = (peer, hash) {
                     self::get::GetInteractive {
                         hash,
-                        opts: iroh_bytes::get::Options {
+                        opts: iroh::bytes::get::Options {
                             addrs,
                             peer_id: peer,
                             keylog: self.keylog,
@@ -290,7 +290,7 @@ pub fn create_quinn_client(
 #[cfg(feature = "metrics")]
 pub fn init_metrics_collection(
     metrics_addr: Option<SocketAddr>,
-    rt: &iroh_bytes::util::runtime::Handle,
+    rt: &iroh::bytes::util::runtime::Handle,
 ) -> Option<tokio::task::JoinHandle<()>> {
     use iroh_metrics::core::Metric;
